requirements.txt
django 
djangorestframework
django-filter
django-extensions
django-clickhouse-backend
clickhouse-driver 
django-clickhouse
clickhouse-connect
apache-airflow